requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
environs==7.2.0
future==0.18.2
idna==2.8
kiwisolver==1.1.0
marshmallow==3.4.0
matplotlib==3.1.3
numpy==1.18.1
pathspec==0.7.0
peewee==3.13.1
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.11.0
python-telegram-bot==12.4.2
regex==2020.1.8
requests==2.22.0
six==1.14.0
toml==0.10.0
tornado==6.0.3
typed-ast==1.4.1
urllib3==1.25.8
